Перенес создание алертов в презентер
