+++ FakeNFT/Catalog/CatalogViewController.swift
@@ -38,30 +38,12 @@ final class CatalogViewController: UIViewController, CatalogViewControlledProtoc
         } completion: { _ in }
     }
 
-    @objc private func sortTapped() {
-        let sortTitle = NSLocalizedString("catalogView.sortTitle",
-                                          comment: "Text displayed like sort alert description")
-        let sortName = NSLocalizedString("catalogView.sortName",
-                                         comment: "Text displayed like sort alert description")
-        let sortNFT = NSLocalizedString("catalogView.sortNFT",
-                                        comment: "Text displayed like sort alert description")
-        let sortClose = NSLocalizedString("catalogView.sortClose",
-                                          comment: "Text displayed like sort alert description")
-
-        let alert = UIAlertController(
-            title: sortTitle,
-            message: nil,
-            preferredStyle: .actionSheet
-        )
-
-        [UIAlertAction(title: sortName, style: UIAlertAction.Style.default) {_ in },
-         UIAlertAction(title: sortNFT, style: UIAlertAction.Style.default) {_ in },
-         UIAlertAction(title: sortClose, style: UIAlertAction.Style.cancel) {_ in }
-        ].forEach{
-            alert.addAction($0)
-        }
-
+    func showAlert(alert: UIAlertController) {
         present(alert, animated: true)
+    }
+
+    @objc private func sortTapped() {
+        presenter.showSortAlert()
     }
 
     private func drawSelf() {
